Refactoring multi-feature modules from dynamic features to libraries
Reverts dynamic feature modules to allow easier implementation of uploading application to 3rd party stores like F-droid

+++ app-data/src/main/kotlin/co/anitrend/data/model/core/airing/AiringSchedule.kt
@@ -30,8 +30,8 @@ data class AiringSchedule(
     val media: Media?,
     @PrimaryKey
     override val id: Long,
-    override val airingAt: Int,
+    override val airingAt: Long,
     override val episode: Int,
     override val mediaId: Long,
-    override val timeUntilAiring: Int
+    override val timeUntilAiring: Long
 ) : IAiringSchedule
